+++ src/main/java/com/cedarsoftware/util/io/JsonValue.java
@@ -1,0 +1,4 @@
+package com.cedarsoftware.util.io;
+
+interface JsonValue {
+}