+++ src/main/java/com/cedarsoftware/util/io/Resolver.java
@@ -1,8 +1,5 @@
 package com.cedarsoftware.util.io;
 
-import static com.cedarsoftware.util.io.JsonObject.ITEMS;
-import static com.cedarsoftware.util.io.JsonObject.KEYS;
-
 import java.lang.reflect.Array;
 import java.lang.reflect.Field;
 import java.util.ArrayDeque;
@@ -19,9 +16,11 @@ import java.util.Optional;
 
 import com.cedarsoftware.util.ReturnType;
 import com.cedarsoftware.util.io.JsonReader.MissingFieldHandler;
-
 import lombok.AccessLevel;
 import lombok.Getter;
+
+import static com.cedarsoftware.util.io.JsonObject.ITEMS;
+import static com.cedarsoftware.util.io.JsonObject.KEYS;
 
 /**
  * This class is used to convert a source of Java Maps that were created from
@@ -333,7 +332,7 @@ public abstract class Resolver implements ReaderContext
      * @return a new Java object of the appropriate type (clazz) using the jsonObj to provide
      * enough hints to get the right class instantiated.  It is not populated when returned.
      */
-    protected Object createInstance(Class clazz, JsonObject jsonObj)
+    protected Object createInstance(Class<?> clazz, JsonObject jsonObj)
     {
         String type = jsonObj.type;